requirements.txt
openai
python-dotenv
aiohttp
pandas
throttler 